Adds the river ID of qsimVis package to the table

River ID can either be defined by verknet data (see details) or a manually entered. The output tables lists the the source of river information and the defined size of the river for plotting as additional columns.

Usage

add_qsimVis_id(aggregated_data, translation_table)

Arguments

aggregated_data

An input dataframe. Either manually entered or created by one of deviating_hours(),adverse_deviation_from_reference() or critical_events()

translation_table

Project specific translation dataframe that contains verknet data and project specific IDs. Column names must be

"ID": Verknet ID or ID of manually added rivers which are also the filnames of the river csv tables

"verknet_BWaStrIdNr": The water body number in verknet. If NA the river information is assumed to be added manually.

"Bezeichnung": The real water body name

"section_name": The project specific ID of the whole river or a river section. Can be multiple IDs per waterbody, separated by ",", however, must be unique.

"river_size": Information about the line width to be drawn from 1 (thin) to 3 (big).

@details The Verknet data is available here: https://www.gdws.wsv.bund.de/DE/service/karten/03_VerkNet-BWaStr/VerkNet-BWaStr_node.html