Adds quality categories to the river table

As a result, each location is linked to a water quality value.

Usage

extend_riverTable(
  rivers,
  river_id,
  aggregated_data,
  varName,
  sixBreaks,
  NA_processing = "interpolation"
)

Arguments

rivers

List of rivers loaded with load_rivers()

river_id

Character with verknet river ID that is one of the names of the rivers list

aggregated_data

A dataframe created by one of deviating_hours(), adverse_deviation_from_reference() or critical_events()

varName

The column name of the agregated data in the output_table

sixBreaks

Breaks defining the lower limits of the categories.

NA_processing

Either "interpolation" or "steps". Defines if NA values between two locations are either interpolated (default) or kept constant based on an upstream value.

Value

The input river_table extended by three columns: 1) "value" containing the exact value, 2) "quality" containing 6 different quality categories based on the defined breaks, 3) "color" assigning th color for plotting

Details

The qsim_misa_table does not provide information for every location within the river, since the local resolution of the qsim output differs from the river course data. Values between two known river sites are interpolated. The color scale cannot be changed. Since in all misa assessment parameters low values refer to a good water quality, the colors range from green for low values to red for high values.