Filters the parameter data frame by time and sites

Filters the loaded data frame by sites and time

Usage

filter_parameter_data(
  dataFrame,
  sites = "",
  tBeg = min(dataFrame$posixDateTime, na.rm = TRUE),
  tEnd = max(dataFrame$posixDateTime, na.rm = TRUE)
)

Arguments

dataFrame

Data frame loaded by a MiSa function (see details)

sites

Vector of strings containing the names of considered sites, as column names of the data frame. If "" (default) all sites are included

tBeg

POSIX-Value with a start time of the observeration interval

tEnd

POSIX-Value with an end time of the observeration interval

Value

A filtered data frame with the same columns as the input data frame

Details

The name of the site column must be "site", the name of the timestamp column must be "posixDateTime".