Reads river data

Reads all csv files starting with "river" in a folder and combines it in a list. Each csv defines a river stretch in the region of interest and must at least contain 5 columns: x: longitude y: lattitude distance_to_neighbour: distance to the previous point of the river in km km: distance from the river estuary qsim_id: The qsim ID as given by function QSIM_prepare() if the point matches a qsim output site

Usage

load_rivers(aggregated_data, translation_table, path_manual = NULL)

Arguments

aggregated_data

A dataframe created by one of deviating_hours(), adverse_deviation_from_reference() or critical_events() with verknet IDs as rownames

translation_table

Project specific translation dataframe that contains verknet data and project specific IDs. Column names must be

"ID": Verknet ID or ID of manually added rivers which are also the filnames of the river csv tables

"verknet_BWaStrIdNr": The water body number in verknet. If NA the river information is assumed to be added manually.

"Bezeichnung": The real water body name

"section_name": The project specific ID of the whole river or a river section. Can be multiple IDs per waterbody, separated by ",", however, must be unique.

"river_size": Information about the line width to be drawn from 1 (thin) to 3 (big).

path_manual

A path to manually added river tables. If NULL (default) no manually added rivers are assumed.